requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-auth==2.23.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
python-socketio==5.10.0 